商品详情图片和描述的显示

+++ src/cn/ucai/FuLiCenter/bean/NewGoodBean.java
@@ -72,7 +72,7 @@ public class NewGoodBean implements Serializable {
     private String colorName;
     private String colorCode;
     private String colorUrl;
-    private int addTime;
+    private Long addTime;
 
     public int getId() {
         return id;
@@ -210,11 +210,11 @@ public class NewGoodBean implements Serializable {
         this.colorUrl = colorUrl;
     }
 
-    public int getAddTime() {
+    public Long getAddTime() {
         return addTime;
     }
 
-    public void setAddTime(int addTime) {
+    public void setAddTime(Long addTime) {
         this.addTime = addTime;
     }
 }
